Improved Anchor Placement Strategy
Carriers will now consider other locations than the nearest well or a well that they're getting stuck on. Also, if a carrier is on low health, it will remove itself from the carrier with anchor on the shared array so that dying with a anchor doesn't prevent future anchors from being made.